+++ cli/src/commands/git/push.rs
@@ -30,6 +30,7 @@ use jj_lib::git;
 use jj_lib::git::GitBranchPushTargets;
 use jj_lib::object_id::ObjectId;
 use jj_lib::op_store::RefTarget;
+use jj_lib::op_store::RemoteRef;
 use jj_lib::refs::classify_bookmark_push_action;
 use jj_lib::refs::BookmarkPushAction;
 use jj_lib::refs::BookmarkPushUpdate;
@@ -48,6 +49,7 @@ use crate::cli_util::RevisionArg;
 use crate::cli_util::WorkspaceCommandHelper;
 use crate::cli_util::WorkspaceCommandTransaction;
 use crate::command_error::user_error;
+use crate::command_error::user_error_with_hint;
 use crate::command_error::CommandError;
 use crate::commands::git::get_single_remote;
 use crate::complete;
@@ -78,7 +80,7 @@ use crate::ui::Ui;
 ///     https://jj-vcs.github.io/jj/latest/bookmarks/#conflicts
 
 #[derive(clap::Args, Clone, Debug)]
-#[command(group(ArgGroup::new("specific").args(&["bookmark", "change", "revisions"]).multiple(true)))]
+#[command(group(ArgGroup::new("specific").args(&["bookmark", "change", "revisions", "named"]).multiple(true)))]
 #[command(group(ArgGroup::new("what").args(&["all", "deleted", "tracked"]).conflicts_with("specific")))]
 pub struct GitPushArgs {
     /// The remote to push to (only named remotes are supported)
@@ -167,6 +169,13 @@ pub struct GitPushArgs {
         add = ArgValueCandidates::new(complete::mutable_revisions)
     )]
     change: Vec<RevisionArg>,
+    /// Specify a new bookmark name and a revision to push under that name, e.g.
+    /// '--named myfeature=@'.
+    ///
+    /// Does not require --allow-new
+    // TODO: Add arg completer
+    #[arg(long, value_name = "NAME=REVISION")]
+    named: Vec<String>,
     /// Only display what will change on the remote
     #[arg(long)]
     dry_run: bool,
@@ -255,7 +264,6 @@ pub fn cmd_git_push(
             };
             (bookmark_name.as_ref(), targets)
         });
-        let view = tx.repo().view();
         for (bookmark_name, targets) in change_bookmarks {
             if !seen_bookmarks.insert(bookmark_name) {
                 continue;
@@ -271,6 +279,77 @@ pub fn cmd_git_push(
             }
         }
 
+        for name_revision in &args.named {
+            let Some((name, revision_str)) = name_revision.split_once('=') else {
+                return Err(user_error_with_hint(
+                    format!(
+                        "Argument '{name_revision}' must include '=' and have the form \
+                         NAME=REVISION"
+                    ),
+                    "For example, `--named myfeature=@` is valid syntax",
+                ));
+            };
+            // Revisions are lenient towards spaces, so let's make the name lenient as well
+            // Git does not allow bookmark names that include spaces anyway.
+            let name = name.trim();
+            if name.is_empty() || revision_str.is_empty() {
+                return Err(user_error_with_hint(
+                    format!(
+                        "Argument '{name_revision}' must have the form NAME=REVISION, with both \
+                         NAME and REVISION non-empty."
+                    ),
+                    "For example, `--named myfeature=@` is valid syntax",
+                ));
+            }
+            // The logic here is similar to but not quite identical to the logic in `jj
+            // bookmark create`
+            if tx.repo().view().get_local_bookmark(name).is_present() {
+                return Err(user_error_with_hint(
+                    format!("Bookmark already exists: '{name}'"),
+                    format!(
+                        "Use 'jj bookmark move' to move it, and 'jj git push -b {name} \
+                         [--allow-new]' to push it."
+                    ),
+                ));
+            }
+            if tx
+                .repo()
+                .view()
+                .get_remote_bookmark(name, &remote)
+                .is_present()
+            {
+                return Err(user_error_with_hint(
+                    format!("Remote bookmark already exists: '{name}@{remote}'"),
+                    "Use 'jj bookmark track' to track it.",
+                ));
+            }
+            let revision = tx
+                .base_workspace_helper()
+                .resolve_single_rev(ui, &revision_str.to_string().into())?;
+            let ref_target = RefTarget::normal(revision.id().clone());
+            tx.repo_mut()
+                .set_local_bookmark_target(name, ref_target.clone());
+
+            let allow_new = true;
+            match classify_bookmark_update(
+                name,
+                &remote,
+                LocalAndRemoteRef {
+                    local_target: &ref_target,
+                    remote_ref: &RemoteRef::absent(),
+                },
+                allow_new,
+            ) {
+                Ok(Some(update)) => bookmark_updates.push((name.to_owned(), update)),
+                Ok(None) => writeln!(
+                    ui.status(),
+                    "Bookmark {name}@{remote} already matches the desired revision",
+                )?,
+                Err(reason) => return Err(reason.into()),
+            }
+        }
+
+        let view = tx.repo().view();
         let allow_new = args.allow_new || tx.settings().get("git.push-new-bookmarks")?;
         let bookmarks_by_name = find_bookmarks_to_push(view, &args.bookmark, &remote)?;
         for &(bookmark_name, targets) in &bookmarks_by_name {
@@ -287,8 +366,10 @@ pub fn cmd_git_push(
             }
         }
 
-        let use_default_revset =
-            args.bookmark.is_empty() && args.change.is_empty() && args.revisions.is_empty();
+        let use_default_revset = args.bookmark.is_empty()
+            && args.change.is_empty()
+            && args.revisions.is_empty()
+            && args.named.is_empty();
         let bookmarks_targeted = find_bookmarks_targeted_by_revisions(
             ui,
             tx.base_workspace_helper(),